streamlit_dashboard/requirements.txt
streamlit
psycopg2-binary
python-dotenv
pandas
sqlalchemy-redshift